+++ packages/subgraph/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Removed token call generators and used the ones from `@aragon/sdk-commons-subgraph`
+- Removed `createERC1155TokenCalls`, `createTokenCalls`, `createWrappedTokenCalls` and `createDummyActions` and use the equivalent functions from `@aragon/sdk-commons-subgraph`
 - Used `createDummyAction` from `@aragon/sdk-commons-subgraph`
 
 ## 1.4.0